(PE-28912) Remove reliance on stdlib module
In order not to have to ship the puppetlabs-stdlib module with this module, this copies the Absolutepath, Windowspath, and Unixpath types, as well as the pick function. Since puppet parser functions get cached together, this changes the name to os_patching_pick to avoid conflicts with users who have stdlib installed.