(PE-29275) Add initial_fact_timeout parameter
Because gathering patch information can be slow, especially on Windows, it has been observed that the exec that runs the command to gather the initial fact information may timeout after the default 300 second exec timeout elapses.  This adds a parameter to pe_patch to allow for setting a different timeout, and sets it to 15 minutes by default. The timeout can be disabled entirely by setting it to 0. This does not affect the cron job/scheduled task, just the initial fact cmd exec.