(PE-29468) Fix pe_patch fact when cached catalog is not what we expect
Previously, the fact generation script would fail if it had problems parsing the cached catalog json file, including if a package resource did not contain a parameters hash. This wraps the json parsing in an exception handler that returns an empty string if it runs into trouble. It also validates the parsed hash so that it can attempt to find other version locked packages in the catalog if one of the resources has problems.

This also replaces the use of the puppet_vardir fact, since this is provided by puppetlabs-stdlib, which we do not ship with the module.  Instead, it gets vardir from puppet config print.