(PE-30100) Detect patching failures on Windows
Previously, this was not inspecting the results JSON file to determine if each patch was successfully applied or not. This collects any patches that failed to apply for whatever reason, and fails the task if any were not applied successfully. This also dumps the results json file into the pe_patch_task.log file for debugging purposes.