(PE-29920) Only fetch Windows updates of type 'Software' and allow override of criteria
Previously, the criteria used for searching for Windows updates to install was 'IsInstalled=0 and IsHidden=0'. This caused driver updates to appear in the list of patches to apply. Most people only care about 'Software' type updates (which is most of everything else), and would prefer not to install drivers via this method. This changes the default to IsInstalled=0 and IsHidden=0 and Type='Software'.

Additionally, a pe_patch::windows_update_criteria parameter is added to the class. This allows the user to define their own criteria to search for updates. To support this, the pe_patch_fact_generation.ps1 and pe_patch_groups.ps1 files are now templates, so that we can inject the value of pe_patch::windows_update_criteria into them. Also, pe_patch_groups.ps1 is now laid down by the pe_patch class on the target, rather than being delivered by the patch_server task, and the patch_server task calls this file. This allows us to use the value a user might have chosen for windows_update_criteria, since we don't have a good way of providing a templated file to a task.