(PE-30287,PE-30310) Pass --environment to puppet facts upload and only upload when necessary
The 'puppet facts upload' that happens at the end of the pe_patch_fact_generation script uploads facts with an environment of 'production' by default. For users with nodes not in the production environment, this results in facts showing up with the wrong environment in PuppetDB. This changes the script to use the --environment flag to upload the fact with the correct environment.

Additionally, 'puppet upload facts' was called, even if the fact had not changed. The scripts now include additional logic to only call 'puppet facts upload' if the package_updates files differs from the previous version of it. That is, when there are new packages that need updating. This will help pe_patch work better at scale and cut down on unnecessary fact uploads.

Finally, because this module resides in the base modulepath, we can't serve files up statically using the 'source' attribute, which results in excessive file metadata checks.  Now that both the linux and windows pe_patch_fact_generation files are templated and provided using 'content' instead, we should no longer have this problem.