(PE-30785,PE-30100) Fail patch_server task when Windows update is skipped due to timeout
Currently, when Windows updates are skipped due to reaching the timeout for the whole patching process, the task will still show that patching was successful, and you have to inspect the debug parameter from the task output to determine some updates were actually skipped due to the timeout. This changes the powershell script to have the update result for skipped tasks to have a status of SkippedDuetoTimeout, which the patch_server task will recognize as a failed update and fail the task appropriately.