(PE-30691) Move reboot and fact refresh into functions, add reboot status
Because the group_patching plan will need to know if a node has actually been rebooted or not, this rearranges a few things in the patch_server task. First, a 'was_rebooted' value is added to the output. Second, do_reboot_if_needed is added, which takes what we were previously doing at the end of the task and moves it into a function that returns if a reboot was triggered or not. Third, refreshing the fact at the end of the task was also moved into a function so we can run it on non-Windows nodes before rebooting.