(PE-30691) Don't do last_boot_time if run_health_check is false
This was previously running the pe_patch::last_boot_time task, even if a user had set run_health_check to false, which means we don't need to check last boot time since we aren't waiting for reboots to complete. This just wraps calling that task in a check for run_health_check.