(PE-30864) Move pe_patch_fact_generation.sh script
Because it is problematic for some users to drop the fact generation script into /usr/local/bin, this moves it to /var/cache/pe_patch along with the other pe_patch files. We'll probably want to move these to /opt/puppetlabs/pe_patch at some point, but this is a first step.