(PE-30691) Add was_rebooted to last_run in fact
Since we now populate the run_history file with was_rebooted, which denotes if we actually rebooted the node or not (e.g. when the reboot parameter is set to 'smart', we may or may not reboot after patching), this adds it to the last_run data in the pe_patch fact.