(PE-30961) Reorganize group_patching plan and always do reboot check
The nesting of if/else for different steps of the plan was a bit hard for me to follow, so it has now been rearranged into clear pre-patch, patch, and post-patch steps. Any variables not defined because a section is skipped are now defaulted at the end of the plan, rather than having to keep track of all the variables throughout the plan. This also changes the behavior of the plan to always wait for nodes to come back from reboot, whether you are doing a post-patch health check or not, since just verifying the node comes back after reboot is useful to users, and we'll want to add a post-reboot command hook later.

This also changes the names of some variables and results to be more clear about what they mean.

Also, this fixes an issue where if you had a set of nodes that went through patching, but only a subset of them needed to be restarted, then only those nodes that were restarted would get the post-patching puppet run check, rather than all nodes that were successfully patched.