(PE-30961) Change health_check_service_running to Boolean
While the underlying agent_health task (pulled from the puppet_health_check module) accepts either a boolean or 'running' or 'stopped' for the target_service_running parameter, it's somewhat inconsistent to have this parameter allow both when other parameters are booleans. A bonus for changing this to only a Boolean rather than a Variant is that the parameter becomes a dropdown in the console, rather than a text input field that required the user to type out 'running' or 'true', for example.