(PE-30961) Add post_patching_command to pe_patch class
This adds a post_patching_command parameter to the pe_patch class that behaves in the same way as the pre_patching_command, but runs after patching is complete, but before a node is rebooted (if needed). As with the pre_patching_command, it lays down a file pointing to a script or binary to be run that is executed at the appropriate time. An additional chunk is added to the pe_patch fact, and the patch_server task is modified to run this command. Additionally, running both pre- and post-patching commands has been extracted to a function in the patch_server task, and the output of the command (both stdout and stderr) are printed to the log.