(PE-30961) Add post_reboot_command to group_patching plan
This adds a post_reboot_command parameter to the group_patching plan that behaves similarly to the pre_patching_command and post_patching_command parameters of the pe_patch class. It takes the absolute path to a file to run after a node has rebooted after being patched. Even if a node doesn't reboot because it didn't need to be or reboot was set to never, this command will still run after patching.

Additionally, even if this command fails, it will not fail the plan. The console will show the result of running the command, including any failures, and the user can decide what to do about it. If the command fails, the post-patching puppet run check will still occur.