(PE-31025) Use full path to puppet in pe_patch tasks
Previously, the agent_health task assumed 'puppet' was on the PATH, which is not always a safe assumption. This makes sure the tasks use the full path to puppet, and fall back to just 'puppet' if it can't find the path in the expected location. A separate ticket will further enhance this to find the installed location on Windows rather than relying on this fallback if it isn't in the default location.