(maint) Don't fail plan if wait_until_available fails during reboot check
We don't want the whole group_patching plan to fail if any node doesn't come back from reboot, as we want to still run post-patching checks on good nodes. So this catches errors in wait_until_available. If any node fails to be reachable before timeout, it'll still be marked as 'pending' in the next iteration through the loop, and will show up as a reboot timeout for that node.