(PE-31025,PE-30725) Look up installed path for Puppet on Windows and use Gem to determine if we're on Windows
This modifies the way we determine if the task is running on Windows to use Gem.win_platform?, which is a little more straightforward than the previous way. This also changes the way to find the path to the puppet batch file on Windows to look up where it was installed via the registry key. This allows the task to work on nodes where Puppet was installed to a non-default location.