(PE-30877) Wait for lock file to clear in Windows scripts
If an existing lock file is found, and it contains the PID of a currently active process, rather than throwing an exception immediately, this makes the script wait up to 10 minutes for the lock file to clear. This will allow the patch_server task to run while a fact refresh is in progress and not cause a failure. This is important as some people like to run the patch_server task on a schedule and don't want to see it fail now and then when it happens to step on top of a fact refresh. This will also allow a fact refresh to wait for a currently running patch_server task.