(PE-30966) Use drpm on el-8 platforms
The `deltarpm` package was renamed to `drpm` on el-8 platforms, this
commit fixes an issue with `$manage_delta_rpm` where it would try to
install `deltarpm` on el-8 platforms, which results in an error. It
should now install `drpm` which has the same function but is correctly
named.

This change also excludes Fedora, since that has an osfamily of `RedHat`
but since its version numbers are all greater than 8 and it still has
`deltarpm` available, it continues to use deltarpm.