Update the SUSE PKGS, SECPKGS, and HELDPKGS
The current regex used for SUSE OSes to try and exclude the header lines fails to properly select patches if the repository name actually includes the word Repository (like "Main Update Repository", which is the default main update repository).  There's an additional regex that is used simply to exclude the header separator.  This change uses tail -n +3 to simply skip the header lines altogether and avoids the necessity of two greps.