(PE-31824) Add sequential_patching to group_patching
This adds a new parameter, sequential_patching to the group_patching plan. When true, nodes in the patch group will be patched, rebooted, and the post-reboot script run in sequential order, rather than all at once.

Additionally, the default for the reboot parameter is now "never".

This also moves the logic to wait for reboot into a separate plan, in order to clean up the group_patching plan and standardize how this logic used, whether we are waiting for one node to reboot or many.