(PE-31824) Modify result of 'last' to remove current time
When doing 'last -1 reboot' on RedHat 6/7-based platforms as well as SLES 12, the end of this line will show the current time, since it shows both the start and end time of the session. Since the plan is simply comparing the raw string to see if we've rebooted, this makes it think we've always rebooted, since the end of this string changes every second. This commit slices off the end time, leaving only the login time.