(PE-33237) Remove duplicates from KB lists in pe_patch_fact_generation.ps1
Occasionally, Microsoft will put out multiple packages tied to the same KB (see the ticket for an example). Since this field is intended just to be a list of KBs remedied by the patches to be applied, we can remove any duplicates in this list. The multiple packages for the same KB will still get applied.