(PE-33536) Add support for Amazon Linux 2
Amazon Linux 2 is EL7. Because facter reports the major OS version as 2, pe_patch was treating it like it was pre-RHEL6. This adds checking of the OS name to treat it like EL7. Additionally, the fact generation script now only special-cases needs-restarting on EL6, since we don't support pre-EL6 (which didn't have this app anyway), and this will allow us to handle AL2 as well as any future versions of RHEL-based platforms appropriately.