(PE-33536) Improve handling of yum history info output
Older versions of yum include a 'Transaction performed with:' section, with lines that were matching the regex, so the output of the patch_server task made it look like those packages had been updated too. This attempts to drop everything before the 'Packages Altered' section we care about, and defaults to checking all lines if we can't find it.