(PE-34863) Catch errors in last_boot_time
If you are running the plan with run_health_check set to false, then if any node is not reachable, the entire plan will fail due to not catching the error coming from running the last_boot_time task. This catches that error, so the nodes that fail this are removed from the list for the next step, but the plan can proceed to patch the rest of the nodes.