(PE-4000) Add package_list parameter to patch_server task
Adding new `package_list` parameter to the `patch_server` task. This accepts an `Array[String]` of package names or Windows KB article IDs to be installed.

Validation is performed to ensure each package/KB specified is listed in `package_updates` or `missing_update_kbs` respectively. This is to fail early/gracefully instead of letting the package manager error out.